updated schema

+++ sql/schema/001_blocks.sql
@@ -1,8 +1,8 @@
 -- +goose Up
 CREATE TABLE blocks (
     blockhash VARCHAR(64),
-    confirmations INTEGER,
-    height INTEGER,
+    confirmations BIGINT,
+    height BIGINT,
     version INTEGER,
     version_hex VARCHAR(64),
     merkle_root VARCHAR(64),
@@ -15,9 +15,9 @@ CREATE TABLE blocks (
     ntx INTEGER,
     previous_block_hash VARCHAR(64),
     next_block_hash VARCHAR(64),
-    stripped_size INTEGER,
-    size INTEGER,
-    weight INTEGER
+    stripped_size BIGINT,
+    size BIGINT,
+    weight BIGINT
 );
 
 -- +goose Down
